refactor: hoist common UB related code
