+++ include/lib/ubfree.hpp
@@ -36,10 +36,11 @@
 #include "lib/lang.hpp"
 #include "lib/logger.hpp"
 #include "lib/naming.hpp"
+#include "lib/ubbase.hpp"
 
 /// UBSan is a visitor that collects constraints to ensure that the
 /// execution of a function is free of undefined behavior like overflow .
-class UBSan : protected symir::SymIRVisitor {
+class UBSan : public UBVisitorBase {
 public:
   struct Stats {
     size_t assertedConstraints = 0;
@@ -54,8 +55,8 @@ public:
   };
 
   UBSan(const symir::Funct &fun, std::vector<std::string> execution) :
-      fun(fun), execution(std::move(execution)), tm(std::make_unique<bitwuzla::TermManager>()),
-      coefTerms(), paramTerms() {
+      UBVisitorBase(std::make_unique<bitwuzla::TermManager>(), bitwuzla::Sort()), fun(fun),
+      execution(std::move(execution)) {
     bvSort = tm->mk_bv_sort(32); // 32-bit for normal integer operations
   }
 
@@ -81,7 +82,8 @@ public:
     versions.clear();
     // DON'T clear coefTerms - coefficients must persist across resets!
     // coefTerms.clear();
-    exprStack = std::stack<bitwuzla::Term>();
+    while (!exprStack.empty())
+      exprStack.pop();
     suffixStack = std::stack<std::string>();
     prevBbl = "___entry_bbl";
     currBbl = "";
@@ -91,38 +93,21 @@ public:
   // Collect constraints that can make the execution UB free
   void Collect() { fun.Accept(*this); }
 
-  // Optimize the set of collected constraints to obtain an optimized set
-  void Optimize();
-
-  // Create a versioned variable expression for the given scalar variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version in the version table is incremented and the new version is
-  // used
+  // Override base class methods to track additional context (verbbls)
   bitwuzla::Term CreateScaExpr(const symir::VarDef *var, int version = -1);
-
-  /// Return the name of the loc-th element of the vector variable
-  std::string GetVecElName(const symir::VarDef *var, int loc) const;
-
-  // Create a versioned variable expression for the loc-th element of the vector variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version in the version table is incremented and the new version is used
-  bitwuzla::Term CreateVecElExpr(const symir::VarDef *var, int loc, int version = -1);
-
-  /// Return the name of the field of the struct variable
-  std::string GetStructFieldName(const symir::VarDef *var, const std::string &field) const;
-
-  // Create a versioned variable expression for the field of the struct variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version in the version table is incremented and the new version is used
   bitwuzla::Term
   CreateStructFieldExpr(const symir::VarDef *var, const std::string &field, int version = -1);
-
-  // Create a versioned variable expression using a suffix
   bitwuzla::Term
   CreateVersionedExpr(const symir::VarDef *var, const std::string &suffix, int version = -1);
 
-  // Create a coefficient expression for the given name
-  bitwuzla::Term CreateCoefExpr(const symir::Coef &coef);
+  // Expose other base class methods publicly
+  using UBVisitorBase::CreateCoefExpr;
+  using UBVisitorBase::CreateScaExpr;
+  using UBVisitorBase::CreateStructFieldExpr;
+  using UBVisitorBase::CreateVecElExpr;
+  using UBVisitorBase::CreateVersionedExpr;
+  using UBVisitorBase::GetStructFieldName;
+  using UBVisitorBase::GetVecElName;
 
   // Generate constraints for making the initialization interesting
   void MakeInitInteresting();
@@ -172,15 +157,9 @@ private:
   // Ensure a term is within global bounds (asserted at most once per term).
   void ensureInRange(const bitwuzla::Term &t);
 
-  // Push an expression to the expression stack
-  void pushExpression(bitwuzla::Term expr) { exprStack.push(std::move(expr)); }
-
-  // Pop an expression from the expression stack
-  bitwuzla::Term popExpression() {
-    auto e = exprStack.top();
-    exprStack.pop();
-    return e;
-  }
+  // Expose base class push/pop expression methods
+  using UBVisitorBase::popExpression;
+  using UBVisitorBase::pushExpression;
 
   // Push an element location to the element location stack
   void pushVecElLoc(int loc) { vecElLocStack.push(loc); }
@@ -209,9 +188,7 @@ private:
   const symir::Funct &fun;            // The function that we're analyzing
   std::vector<std::string> execution; // The execution path of the function
 
-  std::unique_ptr<bitwuzla::TermManager> tm; // The term manager for our constraint collecting
-  std::vector<bitwuzla::Term> constraints;   // The constraints generated by this visitor
-  bitwuzla::Sort bvSort;                     // The bitvector sort for 32-bit integers
+  std::vector<bitwuzla::Term> constraints; // The constraints generated by this visitor
 
   // Dedup helpers
   std::unordered_set<uint64_t> constraintIds{};
@@ -225,17 +202,11 @@ private:
   uint64_t boundedAlready_ = 0;
 
   // Context used in collecting UB-free constraints
-  std::stack<bitwuzla::Term> exprStack{}; // The expression stack for evaluating the SymIR program
   std::stack<int> vecElLocStack{}; // The element location stack for evaluating the SymIR program
   std::stack<std::string> suffixStack{}; // The suffix stack for evaluating variable access
   std::string prevBbl = "", currBbl = "",
               nextBbl = ""; // The previous/current/next blocks been/being/to-be evaluated
-  std::map<std::string, int> versions{};        // The SSA version table for each variable
   std::map<std::string, std::string> verbbls{}; // The defined basic block for the current version
-
-  // Term caches to ensure Bitwuzla terms are reused (Bitwuzla doesn't unify by name!)
-  std::map<std::string, bitwuzla::Term> coefTerms{};  // Cache of coefficient terms
-  std::map<std::string, bitwuzla::Term> paramTerms{}; // Cache of parameter terms (version 0 only)
 
   // Flags controlling the behavior of the visitor
   bool enableInterestCoefs = true; // Whether to make coefficients interesting
