+++ include/lib/ubinject.hpp
@@ -30,12 +30,13 @@
 #include <memory>
 #include "lib/lang.hpp"
 #include "lib/logger.hpp"
+#include "lib/ubbase.hpp"
 
 /// IntUBInject injects int-related undefined behaviors (UBs) into a given block of a given program.
 /// It leverages bitvec theory to collect constraints.
-class IntUBInject : protected symir::SymIRVisitor {
+class IntUBInject : public UBVisitorBase {
 public:
-  IntUBInject() : tm(std::make_unique<bitwuzla::TermManager>()), solver(nullptr) {
+  IntUBInject() : UBVisitorBase(std::make_unique<bitwuzla::TermManager>(), bitwuzla::Sort()) {
     bvSort = tm->mk_bv_sort(33); // 33-bit for overflow detection
   }
 
@@ -65,45 +66,22 @@ public:
         exprStack.size()
     );
     versions.clear();
-    varTerms.clear();
+    paramTerms.clear();
     coefTerms.clear();
   }
 
-  // Create a versioned variable expression for the given variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version table is updated and the new version is used
-  bitwuzla::Term CreateScaExpr(const symir::VarDef *var, int version = -1);
-
-  // Get the name of the loc-th element of the vector variable
-  std::string GetVecElName(const symir::VarDef *var, int loc) const;
-
-  // Create a versioned variable expression for the loc-th element of the vector variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version in the version table is incremented and the new version is used
-  bitwuzla::Term CreateVecElExpr(const symir::VarDef *var, int loc, int version = -1);
-
-  /// Return the name of the field of the struct variable
-  std::string GetStructFieldName(const symir::VarDef *var, const std::string &field) const;
-
-  // Create a versioned variable expression for the field of the struct variable
-  // When version==-1, the current version in the version table is used
-  // When version==-2, the version in the version table is incremented and the new version is used
-  bitwuzla::Term
-  CreateStructFieldExpr(const symir::VarDef *var, const std::string &field, int version = -1);
-
-  // Create a coefficient expression for the given name
-  bitwuzla::Term CreateCoefExpr(const symir::Coef &coef);
+  // Expose base class methods publicly
+  using UBVisitorBase::CreateCoefExpr;
+  using UBVisitorBase::CreateScaExpr;
+  using UBVisitorBase::CreateStructFieldExpr;
+  using UBVisitorBase::CreateVecElExpr;
+  using UBVisitorBase::GetStructFieldName;
+  using UBVisitorBase::GetVecElName;
 
 protected:
-  // Push an expression to the expression stack
-  void pushExpression(bitwuzla::Term expr) { exprStack.push(std::move(expr)); }
-
-  // Pop an expression from the expression stack
-  bitwuzla::Term popExpression() {
-    auto e = exprStack.top();
-    exprStack.pop();
-    return e;
-  }
+  // Expose base class push/pop expression methods
+  using UBVisitorBase::popExpression;
+  using UBVisitorBase::pushExpression;
 
 protected:
   void Visit(const symir::VarUse &v) override;
@@ -138,19 +116,11 @@ private:
   }
 
 private:
-  std::unique_ptr<bitwuzla::TermManager> tm;  // The term manager for constraint collecting
   std::unique_ptr<bitwuzla::Bitwuzla> solver; // The solver for checking the constraints
-  bitwuzla::Sort bvSort;                      // 33-bit bitvector sort for overflow detection
 
   std::vector<bitwuzla::Term> constraints;     // The constraints generated by this visitor
   std::vector<bitwuzla::Term> ubConstraints{}; // The UB-specific constraints that ensure UBs
 
-  // Context used in collecting UB-free constraints
-  std::stack<bitwuzla::Term> exprStack{}; // The expression stack for evaluating the SymIR program
-  std::map<std::string, int> versions{};  // The SSA version table for each variable
-  // Bitwuzla does not intern/unify constants by name, so we must cache Terms.
-  std::map<std::string, bitwuzla::Term> varTerms{};
-  std::map<std::string, bitwuzla::Term> coefTerms{};
   const symir::Funct *currentFun = nullptr;
 };
