+++ lib/ubinject.cpp
@@ -29,119 +29,6 @@
 #include <memory>
 #include "global.hpp"
 
-bitwuzla::Term IntUBInject::CreateScaExpr(const symir::VarDef *var, int version) {
-  Assert(var != nullptr, "Cannot create a variable expression for a nullptr variable");
-  const auto &varName = var->GetName();
-  Assert(
-      var->IsScalar(),
-      "Cannot create a scalar expression for a vector %s. Use the other overload instead.",
-      varName.c_str()
-  );
-  Assert(
-      var->GetType() == symir::SymIR::I32,
-      "Variable %s is of type %s, however only 32-bit integer variables are supported for now!",
-      varName.c_str(), symir::SymIR::GetTypeName(var->GetType()).c_str()
-  );
-  if (version == -1) {
-    version = versions[varName];
-  } else if (version == -2) {
-    version = ++versions[varName];
-  }
-  // We used 33-bit bitvectors since we're overflowing 32-bit integers
-  const std::string fullName = varName + "_v" + std::to_string(version);
-  auto it = varTerms.find(fullName);
-  if (it != varTerms.end()) {
-    return it->second;
-  }
-  auto term = tm->mk_const(bvSort, fullName);
-  varTerms.emplace(fullName, term);
-  return term;
-}
-
-std::string IntUBInject::GetVecElName(const symir::VarDef *var, int loc) const {
-  const auto &varName = var->GetName();
-  Assert(
-      var->IsVector(), "The variable %s is not a vector, cannot get element name", varName.c_str()
-  );
-  const auto numEls = var->GetVecNumEls();
-  Assert(
-      loc < numEls, "The element index (%d) is out of bound (%d) for variable %s", loc, numEls,
-      varName.c_str()
-  );
-  return varName + "_el" + std::to_string(loc);
-}
-
-bitwuzla::Term IntUBInject::CreateVecElExpr(const symir::VarDef *var, int loc, int version) {
-  Assert(var != nullptr, "Cannot create a variable expression for a nullptr array variable");
-  Assert(
-      var->IsVector(), "Cannot create a vector element expression for a scalar variable. "
-                       "Use the other overload instead."
-  );
-  const auto eleName = GetVecElName(var, loc);
-  if (version == -1) {
-    version = versions[eleName];
-  } else if (version == -2) {
-    version = ++versions[eleName];
-  }
-  const std::string fullName = eleName + "_v" + std::to_string(version);
-  auto it = varTerms.find(fullName);
-  if (it != varTerms.end()) {
-    return it->second;
-  }
-  auto term = tm->mk_const(bvSort, fullName);
-  varTerms.emplace(fullName, term);
-  return term;
-}
-
-std::string
-IntUBInject::GetStructFieldName(const symir::VarDef *var, const std::string &field) const {
-  return var->GetName() + "_" + field;
-}
-
-bitwuzla::Term IntUBInject::CreateStructFieldExpr(
-    const symir::VarDef *var, const std::string &field, int version
-) {
-  std::string name = GetStructFieldName(var, field);
-  if (version == -1) {
-    version = versions[name];
-  } else if (version == -2) {
-    version = ++versions[name];
-  }
-  const std::string fullName = name + "_v" + std::to_string(version);
-  auto it = varTerms.find(fullName);
-  if (it != varTerms.end()) {
-    return it->second;
-  }
-  auto term = tm->mk_const(bvSort, fullName);
-  varTerms.emplace(fullName, term);
-  return term;
-}
-
-bitwuzla::Term IntUBInject::CreateCoefExpr(const symir::Coef &coef) {
-  Assert(
-      coef.GetType() == symir::SymIR::I32,
-      "Coefficient %s is of type %s, however only 32-bit integer coefficient are supported for "
-      "now!",
-      coef.GetName().c_str(), symir::SymIR::GetTypeName(coef.GetType()).c_str()
-  );
-  // If we've already extracted the value from the model, we use that value,
-  // otherwise it's an uninterpreted constant which the solver needs to figure
-  // out the value for.
-  // We used 33-bit bitvectors since we're overflowing 32-bit integers
-  if (coef.IsSolved()) {
-    return tm->mk_bv_value_int64(bvSort, coef.GetTypedValue<int>());
-  } else {
-    const std::string &name = coef.GetName();
-    auto it = coefTerms.find(name);
-    if (it != coefTerms.end()) {
-      return it->second;
-    }
-    auto term = tm->mk_const(bvSort, name);
-    coefTerms.emplace(name, term);
-    return term;
-  }
-}
-
 std::unique_ptr<symir::Funct> IntUBInject::InjectUBs(const symir::Funct *f, const symir::Block *b) {
   Assert(b != nullptr, "The basic block to inject UBs is a nullptr");
   currentFun = f;
